Refactor and split ComposerModel into its own module
