+++ crates/wysiwyg/src/composer_model/format_nodes.rs
@@ -1,11 +1,37 @@
-use crate::composer_model::{slice, slice_from, slice_to};
+// Copyright 2022 The Matrix.org Foundation C.I.C.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Copyright 2022 The Matrix.org Foundation C.I.C.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::composer_model::base::{slice, slice_from, slice_to};
 use crate::dom::nodes::{ContainerNodeKind, DomNode, TextNode};
 use crate::dom::{
     Dom, DomHandle, DomLocation, MultipleNodesRange, Range, SameNodeRange,
 };
-use crate::{
-    ComposerModel, ComposerUpdate, InlineFormatType, NodeJoiner, UnicodeString,
-};
+use crate::{ComposerModel, ComposerUpdate, InlineFormatType, UnicodeString};
 
 #[derive(Eq, PartialEq, Debug)]
 enum FormatSelectionType {
@@ -277,13 +303,11 @@ where
             self.state.dom.lookup_node(handle.parent_handle())
         {
             if parent.children().len() - handle.index_in_parent() > 1 {
-                NodeJoiner::join_format_node_with_prev(
-                    &mut self.state.dom,
-                    handle.next_sibling(),
-                );
+                self.join_format_node_with_prev(handle.next_sibling());
             }
         }
-        NodeJoiner::join_format_node_with_prev(&mut self.state.dom, handle);
+        // Merge current node with previous if possible
+        self.join_format_node_with_prev(handle);
     }
 }
